requirements.txt
google_api_python_client==2.154.0
protobuf==5.29.1
Requests==2.32.3
